Add global exception filter

+++ src/auth/auth.service.ts
@@ -1,10 +1,10 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { ConflictException, Injectable, UnauthorizedException } from '@nestjs/common';
-import { JwtService } from '@nestjs/jwt';
-import { UsersService } from 'src/users/users.service';
-import * as bcrypt from 'bcrypt';
+import { ConflictException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { JwtService } from '@nestjs/jwt';
+import * as bcrypt from 'bcrypt';
+import { User } from 'src/users/schemas/user.schema';
+import { UsersService } from 'src/users/users.service';
+import { TokensResDto } from './dto/token.dto';
 @Injectable()
 export class AuthService {
 	constructor(
@@ -13,69 +13,63 @@ export class AuthService {
 		private configService: ConfigService
 	) {}
 
-	async signup(email: string, name: string, password: string) {
+	async signup(email: string, name: string, password: string): Promise<TokensResDto> {
 		const existingUser = await this.usersService.findByEmail(email);
 		if (existingUser) {
 			throw new ConflictException('User already exists');
 		}
 		const user = await this.usersService.create(email, name, password);
 
-		const payload = { sub: user._id, email: user.email };
-		const accessToken = this.jwtService.sign(payload, { expiresIn: '15m' });
-		const refreshToken = this.jwtService.sign(payload, { expiresIn: '7d' });
+		const { accessToken, refreshToken } = await this.setTokens(user);
 
 		return {
-			user: { id: user._id, email: user.email, name: user.name },
-			accessToken: accessToken,
-			refreshToken: refreshToken
+			accessToken,
+			refreshToken,
+			user: {
+				id: user._id,
+				email: user.email,
+				name: user.name
+			}
 		};
 	}
 
-	async signin(username: string, password: string) {
+	async signin(username: string, password: string): Promise<TokensResDto> {
 		const user = await this.usersService.validateUser(username, password);
 		if (!user) {
-			throw new UnauthorizedException('Invalid credentials');
+			throw new UnauthorizedException('Incorrect email or password.');
 		}
 
-		const payload = { sub: user._id, email: user.email };
-		const accessToken = this.jwtService.sign(payload, { expiresIn: '15m' });
-
-		const refreshToken = this.jwtService.sign(payload, { expiresIn: '24h' });
+		const { accessToken, refreshToken } = await this.setTokens(user);
 
 		return {
 			user: { id: user._id, email: user.email, name: user.name },
-			accessToken: accessToken,
-			refreshToken: refreshToken
+			accessToken,
+			refreshToken
 		};
 	}
 
-	async refreshTokens(refreshToken: string) {
+	async refreshTokens(oldRefreshToken: string): Promise<TokensResDto> {
 		try {
-			const payload = this.jwtService.verify(refreshToken, {
+			const payload = this.jwtService.verify<{ sub: string }>(oldRefreshToken, {
 				secret: this.configService.get('JWT_REFRESH_SECRET')
 			});
 
 			const user = await this.usersService.findById(payload.sub);
 			if (!user || !user.hashedRefreshToken) throw new UnauthorizedException();
 
-			const isMatch = await bcrypt.compare(refreshToken, user.hashedRefreshToken);
+			const isMatch = await bcrypt.compare(oldRefreshToken, user.hashedRefreshToken);
 			if (!isMatch) throw new UnauthorizedException('Invalid refresh token');
 
-			const newPayload = { sub: user._id };
-			const accessToken = this.jwtService.sign(newPayload, {
-				expiresIn: this.configService.get<string>('JWT_ACCESS_EXPIRATION_MINUTES'),
-				secret: this.configService.get<string>('JWT_ACCESS_SECRET')
-			});
-			const newRefreshToken = this.jwtService.sign(newPayload, {
-				expiresIn: this.configService.get<string>('JWT_REFRESH_EXPIRATION_DAYS'),
-				secret: this.configService.get<string>('JWT_REFRESH_SECRET')
-			});
-
-			await this.usersService.updateRefreshToken(user._id, newRefreshToken);
+			const { accessToken, refreshToken } = await this.setTokens(user);
 
 			return {
-				access_token: accessToken,
-				refresh_token: newRefreshToken
+				accessToken,
+				refreshToken,
+				user: {
+					id: user._id,
+					email: user.email,
+					name: user.name
+				}
 			};
 		} catch (err) {
 			console.error('Refresh token error:', err);
@@ -83,8 +77,30 @@ export class AuthService {
 		}
 	}
 
-	async logout(userId: string) {
+	async logout(userId: string): Promise<{ message: string }> {
+		const user = await this.usersService.findById(userId);
+
+		if (!user) {
+			throw new NotFoundException('User not found');
+		}
+
 		await this.usersService.removeRefreshToken(userId);
-		return { message: 'Logged out successfully' };
+
+		return {
+			message: 'Logged out successfully'
+		};
+	}
+
+	private async setTokens(user: User) {
+		const payload = { sub: user._id, email: user.email };
+		const accessToken = this.jwtService.sign(payload, {
+			expiresIn: this.configService.get<string>('JWT_ACCESS_EXPIRATION_MINUTES')
+		});
+		const refreshToken = this.jwtService.sign(payload, {
+			expiresIn: this.configService.get<string>('JWT_REFRESH_EXPIRATION_DAYS')
+		});
+
+		await this.usersService.updateRefreshToken(user._id, refreshToken);
+		return { accessToken, refreshToken };
 	}
 }
